+++ src/process.ts
@@ -3,14 +3,12 @@ import { type EnvObject, env } from "./env.ts";
 /**
  * An interface that partially shims the Node.js `global.process`.
  */
-export interface Process
-  extends Partial<Omit<typeof globalThis.process, "versions">> {
+export interface Process extends Partial<Omit<typeof globalThis.process, "versions">> {
   env: EnvObject;
   versions: Record<string, string>;
 }
 
-const _process = (globalThis.process ||
-  Object.create(null)) as unknown as Process;
+const _process = (globalThis.process || Object.create(null)) as unknown as Process;
 
 const processShims: Partial<Process> = {
   versions: {},